% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg2psp.R
\name{cutpath}
\alias{cutpath}
\title{Cut a psp segments into smaller segments.}
\usage{
cutpath(datapsp, maxlength)
}
\arguments{
\item{datapsp}{PSP object to be cut}

\item{maxlength}{Maximum length of segments in the returned psp object}
}
\value{
Returns a new psp object with segments length of \code{maxlength} a max.
}
\description{
Cut a psp segments into smaller segments.
}
