% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectset.psp.R
\name{connectset.psp}
\alias{connectset.psp}
\title{Adds marks to connected segments in a psp}
\usage{
connectset.psp(psp, conn.radius = 0, conn.angle = pi/4)
}
\arguments{
\item{psp}{spatstat psp object}

\item{conn.radius}{segments having ends separated than less then conn.radius will be assumed to belong to the same set}

\item{conn.angle}{maximum angle between segments to belong to the same set}
}
\description{
Adds marks to connected segments in a psp
}
